+++ apps/client/src/components/modals/components/UserSelectorBlock.js
@@ -73,6 +73,7 @@ export default function UserSelectorBlock({
           )
         : selectedUsers
     );
+    setIsLoading(false);
   };
 
   return (