+++ apps/client/src/index.js
@@ -11,7 +11,9 @@ import "@styles/GlobalParam.css";
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <Provider store={store}>
-    <BrowserRouter>
+    <BrowserRouter
+      future={{ v7_startTransition: true, v7_relativeSplatPath: true }}
+    >
       <ConnectLine />
       <App />
     </BrowserRouter>